Got all of this working except tooltips. A div vs span, or Foundation issue. Maybe I just need to get an external plugin for tooltips.
All content —
I want all sections (Top header Body text & Highlight)s to center like
the Title Area is within the window
Generally can't change fonts.
Somehow they are changed on the homepage, though.

Top header —
I want the type to start where the Title Area starts and ends
hovers are on line 1698
background: #272727;

Title Area —
I want that title and author to be on 2 lines
I want Filters to right align, stacked, smaller with corresponding dots
next to them, or the text color coded
I want the tooltip's carrot to be on the left side of the Filter and

Annotations —
What happened to the white border I had on it?
I can't change the font!

Main Text —
I can't change the font, or line-height!
I want this entire section to center like the Title Area is